server/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hfi-rag-server"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langchain-community>=0.3",
    "langchain-chroma>=0.1",
    "langgraph>=0.2",
    "chromadb>=0.5",
    "pymupdf4llm>=0.0.17",
    "pymupdf>=1.24",
    "openai>=1.40",
    "tiktoken>=0.7",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    "pillow>=10.0",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "aiofiles>=23.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "fastapi>=0.111",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "rich"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
